# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: BridalCover CRM API
  version: 1.0.0
  description: |
    **REST API for Lead Management in BridalCover CRM system**
    
    This system was developed following the principles of:
    - **Domain-Driven Design (DDD)**
    - **Hexagonal Architecture (Ports & Adapters)**
    - **Clean Architecture**
    - **API-First Design**
    
    ## Main Features
    - Lead registration and management
    - Robust data validation (CNPJ, Email, Phone)
    - Support for multiple contact channels
    - Lead source tracking
    
    ## Authentication
    Currently the API does not require authentication (under development).
    
  contact:
    name: Gustavo Antunes
    email: gustavo@example.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Development Server
  - url: https://api.bridalcover.com
    description: Production Server (future)

tags:
  - name: Leads
    description: Endpoints for Lead (prospects) management

paths:
  /api/v1/leads:
    post:
      tags:
        - Leads
      summary: Register new lead
      description: |
        Registers a new lead (prospect) in the system.
        
        **Validations:**
        - Company name is required (2-200 characters)
        - CNPJ must be valid (if provided)
        - At least one contact method is required (email, phone, website or socialMedia)
        - Leads from WEBSITE source must have email
        - Leads from COLD_CALL source must have phone
      operationId: registerLead
      requestBody:
        description: Lead data to be registered
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadRequest'
            examples:
              completeLead:
                summary: Complete Lead
                value:
                  companyName: Vestidos Elegantes
                  cnpj: "17.233.049/0001-97"
                  email: contato@vestidoselegantes.com.br
                  phone: "11987654321"
                  website: www.vestidoselegantes.com.br
                  socialMedia:
                    INSTAGRAM: "@vestidoselegantes"
                    FACEBOOK: vestidoselegantes
                  source: MANUAL_ENTRY
              simpleLead:
                summary: Simple Lead
                value:
                  companyName: Boutique da Noiva
                  phone: "11987654321"
                  source: COLD_CALL
      responses:
        '201':
          description: Lead successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validationError:
                  summary: Validation Error
                  value:
                    status: 400
                    error: Bad Request
                    message: "Invalid request: Company name cannot be empty"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    get:
      tags:
        - Leads
      summary: List all leads
      description: Retrieves a paginated list of all leads in the system
      operationId: listLeads
      parameters:
        - name: page
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
          example: 0
        - name: size
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          example: 20
      responses:
        '200':
          description: Leads retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedLeadResponse'
        '400':
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/leads/{id}:
    get:
      tags:
        - Leads
      summary: Get lead by ID
      description: Retrieves the complete data of a specific lead by its UUID
      operationId: getLeadById
      parameters:
        - name: id
          in: path
          description: Lead UUID
          required: true
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Lead found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadResponse'
        '400':
          description: Invalid ID (not a valid UUID)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidId:
                  summary: Invalid ID
                  value:
                    status: 400
                    error: Bad Request
                    message: "Invalid lead ID: LeadId must be a valid UUID"
        '404':
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  summary: Lead Not Found
                  value:
                    status: 404
                    error: Not Found
                    message: "Lead not found with ID: 550e8400-e29b-41d4-a716-446655440000"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Leads
      summary: Update an existing lead
      description: |
        Updates an existing lead's information. All fields are optional.
        Only the provided fields will be updated.
      operationId: updateLead
      parameters:
        - name: id
          in: path
          description: Lead UUID
          required: true
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
      requestBody:
        description: Updated lead data (all fields are optional)
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLeadRequest'
            examples:
              updateContactInfo:
                summary: Update Contact Info
                value:
                  email: novo@email.com.br
                  phone: "11999887766"
              updateStatus:
                summary: Update Status
                value:
                  status: CONTACTED
              fullUpdate:
                summary: Full Update
                value:
                  companyName: Novo Nome
                  cnpj: "12.345.678/0001-90"
                  email: novo@email.com.br
                  phone: "11999887766"
                  instagram: "@novonome"
                  status: QUALIFIED
      responses:
        '200':
          description: Lead updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Leads
      summary: Delete a lead
      description: Permanently deletes a lead from the system by its UUID
      operationId: deleteLead
      parameters:
        - name: id
          in: path
          description: Lead UUID
          required: true
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '204':
          description: Lead deleted successfully
        '400':
          description: Invalid ID (not a valid UUID)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    LeadRequest:
      type: object
      description: Data for registering a new lead
      required:
        - companyName
        - source
      properties:
        companyName:
          type: string
          description: Company name
          minLength: 2
          maxLength: 200
          example: Vestidos Elegantes
        cnpj:
          type: string
          description: Brazilian company tax ID (optional, but must be valid if provided)
          pattern: '^\d{2}\.\d{3}\.\d{3}/\d{4}-\d{2}$'
          example: "12.345.678/0001-90"
        email:
          type: string
          format: email
          description: Contact email (optional, but required for WEBSITE source leads)
          example: contato@vestidoselegantes.com.br
        phone:
          type: string
          description: Contact phone (optional, but required for COLD_CALL source leads). Accepts Brazilian formats with or without area code
          example: "11987654321"
        website:
          type: string
          description: Company website (optional)
          example: www.vestidoselegantes.com.br
        socialMedia:
          type: object
          description: Company social media (optional)
          additionalProperties:
            type: string
          example:
            INSTAGRAM: "@vestidoselegantes"
            FACEBOOK: vestidoselegantes
        source:
          type: string
          description: Lead source
          enum:
            - MANUAL_ENTRY
            - GOOGLE_PLACES
            - REFERRAL
            - WEBSITE
            - COLD_CALL
            - SOCIAL_MEDIA
          example: MANUAL_ENTRY

    LeadResponse:
      type: object
      description: Lead response data
      required:
        - id
        - companyName
        - contactInfo
        - status
        - source
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique lead UUID
          example: 550e8400-e29b-41d4-a716-446655440000
        companyName:
          type: string
          description: Company name
          example: Vestidos Elegantes
        cnpj:
          type: string
          description: Brazilian company tax ID (formatted)
          example: "12.345.678/0001-90"
        contactInfo:
          $ref: '#/components/schemas/ContactInfoResponse'
        status:
          type: string
          description: Current lead status
          enum:
            - NEW
            - CONTACTED
            - QUALIFIED
            - PROPOSAL_SENT
            - NEGOTIATING
            - CONVERTED
            - LOST
          example: NEW
        source:
          type: string
          description: Lead source
          enum:
            - MANUAL_ENTRY
            - GOOGLE_PLACES
            - REFERRAL
            - WEBSITE
            - COLD_CALL
            - SOCIAL_MEDIA
          example: MANUAL_ENTRY
        createdAt:
          type: string
          format: date-time
          description: Lead creation date and time
          example: "2025-11-01T10:30:00"
        updatedAt:
          type: string
          format: date-time
          description: Last update date and time
          example: "2025-11-01T10:30:00"

    ContactInfoResponse:
      type: object
      description: Contact information
      required:
        - hasEmail
        - hasPhone
        - isComplete
      properties:
        email:
          type: string
          format: email
          description: Contact email (normalized)
          example: contato@vestidoselegantes.com.br
        phone:
          type: string
          description: Complete phone with country code and area code
          example: "5511987654321"
        phoneFormatted:
          type: string
          description: Formatted phone for display
          example: "(11) 98765-4321"
        website:
          type: string
          description: Company website
          example: www.vestidoselegantes.com.br
        socialMedia:
          type: object
          description: Company social media
          additionalProperties:
            type: string
          example:
            INSTAGRAM: "@vestidoselegantes"
            FACEBOOK: vestidoselegantes
        hasEmail:
          type: boolean
          description: Indicates if has email
          example: true
        hasPhone:
          type: boolean
          description: Indicates if has phone
          example: true
        isComplete:
          type: boolean
          description: Indicates if contact information is complete (email AND phone)
          example: true

    UpdateLeadRequest:
      type: object
      description: Data for updating an existing lead (all fields are optional)
      properties:
        companyName:
          type: string
          description: Updated company name
          minLength: 2
          maxLength: 200
          example: Novo Nome da Empresa
        cnpj:
          type: string
          description: Updated Brazilian company tax ID
          pattern: '^\d{2}\.\d{3}\.\d{3}/\d{4}-\d{2}$'
          example: "12.345.678/0001-90"
        email:
          type: string
          format: email
          description: Updated contact email
          example: novoemail@empresa.com.br
        phone:
          type: string
          description: Updated contact phone
          example: "11999887766"
        instagram:
          type: string
          description: Updated Instagram handle
          example: "@novoinstagram"
        status:
          type: string
          description: Updated lead status
          enum:
            - NEW
            - CONTACTED
            - QUALIFIED
            - PROPOSAL_SENT
            - NEGOTIATING
            - CONVERTED
            - LOST
          example: CONTACTED

    PagedLeadResponse:
      type: object
      description: Paginated response containing leads
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
        - first
        - last
      properties:
        content:
          type: array
          description: Array of leads for the current page
          items:
            $ref: '#/components/schemas/LeadResponse'
        page:
          type: integer
          description: Current page number (0-based)
          example: 0
        size:
          type: integer
          description: Number of items per page
          example: 20
        totalElements:
          type: integer
          format: int64
          description: Total number of leads in the database
          example: 150
        totalPages:
          type: integer
          description: Total number of pages
          example: 8
        first:
          type: boolean
          description: Indicates if this is the first page
          example: true
        last:
          type: boolean
          description: Indicates if this is the last page
          example: false

    ErrorResponse:
      type: object
      description: Standard error response
      required:
        - status
        - error
        - message
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: HTTP error description
          example: Bad Request
        message:
          type: string
          description: Detailed error message
          example: Invalid request data
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        path:
          type: string
          description: Request path that generated the error
          example: /api/v1/leads